+++ src/main/java/racingcar/model/RacingCarMovingCondition.java
@@ -1,0 +1,16 @@
+package racingcar.model;
+
+import camp.nextstep.edu.missionutils.Randoms;
+import racingcar.constants.RacingCarRole;
+
+public class RacingCarMovingCondition implements MovingCondition{
+
+    private static final int MIN_NUMBER = 0;
+    private static final int MAX_NUMBER = 9;
+
+    @Override
+    public boolean canMove() {
+        int number = Randoms.pickNumberInRange(MIN_NUMBER, MAX_NUMBER);
+        return number >= RacingCarRole.MOVING_CONDITION.getValue();
+    }
+}
