+++ src/test/java/racingcar/model/RacingCarMovingConditionTest.java
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("움직이는 조건 객체에 대해 ")
-class MovingConditionTest {
+class RacingCarMovingConditionTest {
 
-    private final MovingCondition movingCondition = new MovingCondition();
+    private final RacingCarMovingCondition movingCondition = new RacingCarMovingCondition();
 
     @Test
     @DisplayName("레이싱 카는 움직일 수도 있고 안 움직일 수도 있다.")
